chore(structured-list): remove next folder
