--- packages/react/src/components/StructuredList/next/StructuredList.mdx
@@ -1,23 +1,0 @@
-import { Props } from '@storybook/addon-docs';
-
-# StructuredList
-
-[Source code](https://github.com/carbon-design-system/carbon/tree/main/packages/react/src/components/StructuredList)
-&nbsp;|&nbsp;
-[Usage guidelines](https://www.carbondesignsystem.com/components/structured-list/usage)
-&nbsp;|&nbsp;
-[Accessibility](https://www.carbondesignsystem.com/components/structured-list/accessibility)
-
-## Table of Contents
-
-## Overview
-
-## Component API
-
-<Props />
-
-## Feedback
-
-Help us improve this component by providing feedback, asking questions on Slack,
-or updating this file on
-[GitHub](https://github.com/carbon-design-system/carbon/edit/main/packages/react/src/components/StructuredList/StructuredList.mdx).
